learnコマンド、初回、rmseの計算で落ちることがあるのを修正。(まだ修正出来てないかも。よくわからない)

+++ source/misc.h
@@ -193,7 +193,7 @@ inline uint64_t get_thread_id()
 // UniformRandomNumberGenerator互換にして、std::shuffle()等でも使えるようにするべきか？
 struct PRNG
 {
-	PRNG(uint64_t seed) : s(seed) { ASSERT_LV1(seed); }
+	PRNG(u64 seed) : s(seed) { ASSERT_LV1(seed); }
 
 	// 時刻などでseedを初期化する。
 	PRNG() {
@@ -210,11 +210,11 @@ struct PRNG
 	template<typename T> T rand() { return T(rand64()); }
 
 	// 0からn-1までの乱数を返す。(一様分布ではないが現実的にはこれで十分)
-	uint64_t rand(size_t n) { return rand<uint64_t>() % n; }
+	u64 rand(u64 n) { return rand<u64>() % n; }
 
 private:
-	uint64_t s;
-	uint64_t rand64() {
+	u64 s;
+	u64 rand64() {
 		s ^= s >> 12, s ^= s << 25, s ^= s >> 27;
 		return s * 2685821657736338717LL;
 	}
